Modified SNPAssignmentFeatureExtractor to operate when no speaker_id field is available.
This can be the situation when a user is using a corpus that AuToBI does not know how to extract a speaker_id field for when loading the regions from the TextGrid or .ala files.

Under this change, the default behavior is to load the same SpeakerNormalizationParameters object specified through the -normalization_parameters flag to all regions.

To use the previous behavior, relying on the readers to set the speaker_id field, the flag -known_speakers must be set.

This is a medium term fix.  It would be preferrable to allow a user the ability to set a speaker_id field more easily.  Using command line parameters to specify a regular expression used to extract a speaker id from a filename would fit with the current operations for BDC and BURNC material.  Eventually it may be valuable to change the single wav file-single speaker requirement, and allow speaker meta information to be tagged through a TextGrid tier, or other file.