Major memory improvement.
required features as defined by a FeatureSet are not removed from a region for the duration of the training and testing of a classifier.  After this change, these features are stored in a list associated with each Region rather than a <String, Object> hash.  This allows for the elimination of overhead to store the String variables holding the attribute names for all of the required features in a FeatureSet.  All of the other features are removed after the extractFeatures(...) routines are all called.  This reduced the memory requirements for training by approximately 33%.