Implemented RAPT pitch extraction algorithm.
This is in service of moving AuToBI out of GPL licensing, and into a BSD-style license.

Praat is released under GPL, and therefore the (very derivative) port of the Sound to Pitch... in PitchExtractor is also GPL.
ESPS and the SPTK implementation of RAPT (aka esps getF0()) are BSD style.  The implementation used in RAPTPitchExtractor is a port of the SPTK implementation (which in turn includes the ESPS implementation nearly verbatim).

PitchExtractor, and associated Praat derived code, is still included here, though deprecated. The license change will come once all GPL dependencies are removed. A this point all derivative code will be removed from the project.

(Also cleaned up some unused imports and removed some finished TODOs)