Implemented Intensity extraction algorithm based on http://www.fon.hum.uva.nl/praat/manual/Sound__Get_intensity__dB_.html.
This is in service of moving AuToBI out of GPL licensing, and into a BSD-style license.

Praat is released under GPL, and therefore the (very derivative) port of the Sound to Intensity... in the previous IntensityExtractor was also GPL.

So I wrote this from scratch based on the praat documentation in the manual rather than reference to the code.  The results are a minorly different -- I used a hanning window rather than a gaussian window, and don't determine the analysis window or frame size based on a minimum pitch value -- but the values are very close.